pyproject.toml
[tool.poetry]
name = "ui-api-practice"
version = "0.1.0"
description = "UI Automation Practice Sprint 3"
authors = ["Igor <igorpushkarev22@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
pytest = "^8.3"
python-dotenv = "^1.0"
faker = "^37.1"
requests = "^2.33.1"
pydantic = "^2.13.3"
pytest-mock = "^3.15.1"
responses = "^0.26.0"
allure-pytest = "^2.16.0"
flake8 = "^7.3.0"
isort = "^8.0.1"
mypy = "^1.20.2"
types-requests = "^2.33.0.20260408"

[tool.poetry.group.playwright.dependencies]
playwright = "^1.48"
pytest-playwright = "^0.7"

[tool.poetry.group.selenium.dependencies]
selenium = "^4.31"
webdriver-manager = "^4.0"
selene = "^1.0.2"

[tool.poetry.group.dev.dependencies]
allure-pytest = "^2.13"
black = "^24.8"
ruff = "^0.6"
isort = "^8.0.1"
flake8-pyproject = "^1.2.4"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
